requirements.txt
fastapi-utils==0.2.1
fastapi==0.104.0
fire==0.5.0
gunicorn==21.2.0
pandas==2.1.1
psycopg2-binary==2.9.9
PySnooper==1.2.0
python-dotenv==1.0.0
requests==2.31.0
sqlmodel==0.0.8
uvicorn==0.23.2